o [NEXUS-4882] Guard against null value

+++ nexus/nexus-test-harness/nexus-test-harness-its/src/test/java/org/sonatype/nexus/integrationtests/nexus4579/Nexus4579OptionalTrashForSnapshotsIT.java
@@ -103,7 +103,11 @@ public class Nexus4579OptionalTrashForSnapshotsIT
         throws IOException
     {
         // clean trash before tests to remove possible leftovers from previous tests
-        FileUtils.deleteDirectory( trashPath );
+        // trashPath may be null if super.@Before failed
+        if ( trashPath != null )
+        {
+            FileUtils.deleteDirectory( trashPath );
+        }
     }
 
     @Test
